backend/requirements.txt
astroid==2.3.1
autopep8==1.4.4
colorama==0.4.1
coverage==4.5.4
Django==2.2.6
django-cors-headers==3.1.1
django-seed==0.1.9
djangorestframework==3.10.3
Faker==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.2.0
pycodestyle==2.5.0
pylint==2.4.2
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.0
wrapt==1.11.2
